+++ src/main/java/com/moments/models/Moment.java
@@ -7,11 +7,11 @@ import java.util.Date;
 public class Moment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String body;
 
-    private Integer userId;
+    private Long userId;
 
     @Temporal(TemporalType.DATE)
     private Date createdAt;
@@ -21,18 +21,18 @@ public class Moment {
 
     public Moment() {}
 
-    public Moment(String body, Integer userId) {
+    public Moment(String body, Long userId) {
         this.body = body;
         this.userId = userId;
         createdAt = new Date();
         updatedAt = new Date();
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -44,11 +44,11 @@ public class Moment {
         this.body = body;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
