+++ src/main/java/com/moments/services/followings/FollowingService.java
@@ -1,5 +1,5 @@
 package com.moments.services.followings;
 
 public interface FollowingService {
-    void followingEachOther(Integer userId, Integer followerId) throws AlreadyFollowingException;
+    void followingEachOther(Long userId, Long followerId) throws AlreadyFollowingException;
 }
