+++ src/main/java/com/moments/services/moments/MomentService.java
@@ -6,5 +6,5 @@ import com.moments.models.User;
 import java.util.List;
 
 public interface MomentService {
-    List<Moment> findMomentsOfUser(User user, Integer lastMomentId);
+    List<Moment> findMomentsOfUser(User user, Long lastMomentId);
 }
